# Speed up fetchmany() and fetchall()