Optimized the code that transposes columnar batches to rows, now it
takes similar time as the related logic in impala-shell.

Benchmarked with "select * from tpch_parquet.lineitem limit 100000"
(convert_types=False, batches of 1024):
Python 2: 1.38 sec -> 0.89 sec
Python 3: 1.25 sec -> 0.76 sec

Testing:
- Ran the tests with Python 2 and 3, ImpalaDBAPI20Test gives fair
  coverage to the modified functions.
- Also tested row based batches by using HIVE_CLI_SERVICE_PROTOCOL_V5.
  There were no new broken tests besides the ones that already fail.